@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -14,9 +15,11 @@ namespace Calendar\Controller;
 
 use Calendar\Entity\Type;
 use Calendar\Service\CalendarService;
-use Zend\Http\Response;
-use Zend\Mvc\Controller\AbstractActionController;
+use Laminas\Http\Response;
+use Laminas\Mvc\Controller\AbstractActionController;
 use ZfcTwig\View\TwigRenderer;
+
+use function file_put_contents;
 
 /**
  * Class CalendarController
@@ -25,14 +28,8 @@ use ZfcTwig\View\TwigRenderer;
  */
 final class CalendarController extends AbstractActionController
 {
-    /**
-     * @var CalendarService
-     */
-    private $calendarService;
-    /**
-     * @var TwigRenderer
-     */
-    private $renderer;
+    private CalendarService $calendarService;
+    private TwigRenderer $renderer;
 
     public function __construct(CalendarService $calendarService, TwigRenderer $renderer)
     {
@@ -53,7 +50,7 @@ final class CalendarController extends AbstractActionController
             ]
         );
         //Save a copy of the file in the caching-folder
-        \file_put_contents($cacheFileName, $css);
+        file_put_contents($cacheFileName, $css);
 
         /** @var Response $response */
         $response = $this->getResponse();
