@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category   Program
  *
  * @author     Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright  Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright  Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license    https://itea3.org/license.txt proprietary
  *
  * @link       https://itea3.org
@@ -19,36 +20,24 @@ use Calendar\Entity\Calendar;
 use Calendar\Entity\Contact as CalendarContact;
 use Calendar\Options\ModuleOptions;
 use Calendar\Service\CalendarService;
-use Zend\Mvc\Controller\Plugin\AbstractPlugin;
+use Laminas\Mvc\Controller\Plugin\AbstractPlugin;
 use ZfcTwig\View\TwigRenderer;
+
+use function array_chunk;
+use function count;
+use function max;
 
 /**
  * Class RenderCalendarContactList
  *
  * @package Calendar\Controller\Plugin
  */
-class RenderCalendarContactList extends AbstractPlugin
+final class RenderCalendarContactList extends AbstractPlugin
 {
-    /**
-     * @var TwigRenderer
-     */
-    protected $twigRenderer;
-    /**
-     * @var ModuleOptions
-     */
-    protected $moduleOptions;
-    /**
-     * @var CalendarService
-     */
-    protected $calendarService;
+    private TwigRenderer $twigRenderer;
+    private ModuleOptions $moduleOptions;
+    private CalendarService $calendarService;
 
-    /**
-     * RenderCalendarContactList constructor.
-     *
-     * @param TwigRenderer    $twigRenderer
-     * @param ModuleOptions   $moduleOptions
-     * @param CalendarService $calendarService
-     */
     public function __construct(
         TwigRenderer $twigRenderer,
         ModuleOptions $moduleOptions,
@@ -59,12 +48,6 @@ class RenderCalendarContactList extends AbstractPlugin
         $this->calendarService = $calendarService;
     }
 
-
-    /**
-     * @param Calendar $calendar
-     *
-     * @return CalendarPdf
-     */
     public function renderPresenceList(Calendar $calendar): CalendarPdf
     {
         $pdf = new CalendarPdf();
@@ -80,12 +63,9 @@ class RenderCalendarContactList extends AbstractPlugin
 
         //Create chunks of arrays per 22, as that amount fits on the screen
         $paginatedContacts = array_chunk($calendarContacts, 22);
-        $minAmountOfPages = max(\count($paginatedContacts), 1);
+        $minAmountOfPages = max(count($paginatedContacts), 1);
 
         for ($i = 0; $i < $minAmountOfPages; $i++) {
-            /**
-             *
-             */
             $contactListContent = $this->twigRenderer->render(
                 'calendar/pdf/presence-list',
                 [
@@ -108,11 +88,6 @@ class RenderCalendarContactList extends AbstractPlugin
         return $pdf;
     }
 
-    /**
-     * @param Calendar $calendar
-     *
-     * @return CalendarPdf
-     */
     public function renderSignatureList(Calendar $calendar): CalendarPdf
     {
         $pdf = new CalendarPdf();
@@ -127,7 +102,7 @@ class RenderCalendarContactList extends AbstractPlugin
 
         //Create chunks of arrays per 13, as that amount fits on the screen
         $paginatedContacts = array_chunk($calendarContacts, 13);
-        $minAmountOfPages = max(\count($paginatedContacts), 2);
+        $minAmountOfPages = max(count($paginatedContacts), 2);
 
         for ($i = 0; $i < $minAmountOfPages; $i++) {
             /*
