@@ -1,13 +1,6 @@
 <?php
-/**
- * ITEA Office copyright message placeholder
- *
- * @category    Calendar
- * @package     Config
- * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
- */
 
+use BjyAuthorize\Guard\Route;
 use Calendar\Acl\Assertion\Calendar as CalendarAssertion;
 use Calendar\Acl\Assertion\Contact as ContactAssertion;
 use Calendar\Acl\Assertion\Document as DocumentAssertion;
@@ -20,7 +13,7 @@ return [
             /* If this guard is specified here (i.e. it is enabled], it will block
              * access to all routes unless they are specified here.
              */
-            'BjyAuthorize\Guard\Route' => [
+            Route::class => [
                 ['route' => 'assets/calendar-type-color-css', 'roles' => []],
                 [
                     'route' => 'community/calendar/overview',
@@ -28,84 +21,73 @@ return [
                 ],
                 [
                     'route'     => 'community/calendar/calendar',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/review-calendar',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/download-review-calendar',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/select-attendees',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/send-message',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/download-binder',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/presence-list',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/signature-list',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => CalendarAssertion::class,
                 ],
                 [
-                    'route'     => 'community/calendar/update-status',
-                    'roles'     => ['user'],
+                    'route'     => 'json/calendar/update-status',
+                    'roles'     => [],
                     'assertion' => ContactAssertion::class,
                 ],
                 [
                     'route' => 'community/calendar/contact',
-                    'roles' => ['user'],
+                    'roles' => [],
                 ],
                 [
                     'route'     => 'community/calendar/document/document',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => DocumentAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/document/download',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => DocumentAssertion::class,
                 ],
                 [
                     'route'     => 'community/calendar/document/edit',
-                    'roles'     => ['user'],
+                    'roles'     => [],
                     'assertion' => DocumentAssertion::class,
                 ],
-                [
-                    'route' => 'zfcadmin/calendar/overview',
-                    'roles' => ['office'],
-                ],
-                [
-                    'route' => 'zfcadmin/calendar/edit',
-                    'roles' => ['office'],
-                ],
-                [
-                    'route' => 'zfcadmin/calendar/calendar',
-                    'roles' => ['office'],
-                ],
-                [
-                    'route' => 'zfcadmin/calendar/new',
-                    'roles' => ['office'],
-                ],
+                ['route' => 'zfcadmin/calendar/overview', 'roles' => ['office'],],
+                ['route' => 'zfcadmin/calendar/edit', 'roles' => ['office'],],
+                ['route' => 'zfcadmin/calendar/calendar', 'roles' => ['office'],],
+                ['route' => 'zfcadmin/calendar/new', 'roles' => ['office'],],
+                ['route' => 'zfcadmin/calendar/add-contact', 'roles' => ['office'],],
                 [
                     'route' => 'zfcadmin/calendar/select-attendees',
                     'roles' => ['office'],
@@ -126,6 +108,10 @@ return [
                     'route' => 'zfcadmin/calendar/document/edit',
                     'roles' => ['office'],
                 ],
+                ['route' => 'zfcadmin/calendar/type/edit', 'roles' => ['office'],],
+                ['route' => 'zfcadmin/calendar/type/view', 'roles' => ['office'],],
+                ['route' => 'zfcadmin/calendar/type/new', 'roles' => ['office'],],
+                ['route' => 'zfcadmin/calendar/type/list', 'roles' => ['office'],],
             ],
         ],
     ],
