@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Jield BV all rights reserved
  *
@@ -23,10 +24,16 @@ use Contact\Entity\Contact;
 use Contact\Service\ContactService;
 use Doctrine\ORM\PersistentCollection;
 use Interop\Container\ContainerInterface;
-use Zend\Authentication\AuthenticationService;
-use Zend\Http\PhpEnvironment\Request;
-use Zend\Permissions\Acl\Assertion\AssertionInterface;
-use Zend\Router\Http\RouteMatch;
+use Laminas\Authentication\AuthenticationService;
+use Laminas\Http\PhpEnvironment\Request;
+use Laminas\Permissions\Acl\Assertion\AssertionInterface;
+use Laminas\Router\Http\RouteMatch;
+
+use function count;
+use function in_array;
+use function is_array;
+use function strpos;
+use function strtolower;
 
 /**
  * Class AbstractAssertion
@@ -79,12 +86,9 @@ abstract class AbstractAssertion implements AssertionInterface
 
     public function routeHasString(string $string): bool
     {
-        return $this->hasRouteMatch() && \strpos($this->getRouteMatch()->getMatchedRouteName(), $string) !== false;
+        return $this->hasRouteMatch() && strpos($this->getRouteMatch()->getMatchedRouteName(), $string) !== false;
     }
 
-    /**
-     * @return bool
-     */
     public function hasRouteMatch(): bool
     {
         return null !== $this->getRouteMatch()->getMatchedRouteName();
@@ -100,14 +104,6 @@ abstract class AbstractAssertion implements AssertionInterface
         return new RouteMatch([]);
     }
 
-    protected function getRequest(): Request
-    {
-        return $this->container->get('Application')->getMvcEvent()->getRequest();
-    }
-
-    /**
-     * @return string
-     */
     public function getPrivilege(): string
     {
         /**
@@ -135,7 +131,7 @@ abstract class AbstractAssertion implements AssertionInterface
         if (null !== $this->getRequest()->getPost('id')) {
             return (int)$this->getRequest()->getPost('id');
         }
-        if (!$this->hasRouteMatch()) {
+        if (! $this->hasRouteMatch()) {
             return null;
         }
         if (null !== $this->getRouteMatch()->getParam('id')) {
@@ -145,20 +141,28 @@ abstract class AbstractAssertion implements AssertionInterface
         return null;
     }
 
+    protected function getRequest(): Request
+    {
+        return $this->container->get('Application')->getMvcEvent()->getRequest();
+    }
+
     public function rolesHaveAccess($accessRoleOrCollection): bool
     {
         $accessRoles = $this->prepareAccessRoles($accessRoleOrCollection);
-        if (\count($accessRoles) === 0) {
+        if (count($accessRoles) === 0) {
             return true;
         }
 
         foreach ($accessRoles as $access) {
-            if ($access === strtolower(Access::ACCESS_PUBLIC)) {
+            $accessNormalised = strtolower((string)$access);
+
+            if ($accessNormalised === strtolower(Access::ACCESS_PUBLIC)) {
                 return true;
             }
-            if ($this->hasContact()
-                && \in_array(
-                    $access,
+            if (
+                $this->hasContact()
+                && in_array(
+                    $accessNormalised,
                     $this->adminService->findAccessRolesByContactAsArray($this->contact),
                     true
                 )
@@ -172,11 +176,11 @@ abstract class AbstractAssertion implements AssertionInterface
 
     private function prepareAccessRoles($accessRoleOrCollection): array
     {
-        if (!$accessRoleOrCollection instanceof PersistentCollection) {
+        if (! $accessRoleOrCollection instanceof PersistentCollection) {
             /*
              * We only have a string or array, so we need to lookup the role
              */
-            if (\is_array($accessRoleOrCollection)) {
+            if (is_array($accessRoleOrCollection)) {
                 foreach ($accessRoleOrCollection as $key => $accessItem) {
                     $access = $this->adminService->findAccessByName($accessItem);
 
