@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -14,26 +15,12 @@ namespace Calendar\Acl\Assertion;
 
 use Admin\Entity\Access;
 use Calendar\Entity\Calendar as CalendarEntity;
-use Zend\Permissions\Acl\Acl;
-use Zend\Permissions\Acl\Resource\ResourceInterface;
-use Zend\Permissions\Acl\Role\RoleInterface;
+use Laminas\Permissions\Acl\Acl;
+use Laminas\Permissions\Acl\Resource\ResourceInterface;
+use Laminas\Permissions\Acl\Role\RoleInterface;
 
-class Calendar extends AbstractAssertion
+final class Calendar extends AbstractAssertion
 {
-    /**
-     * Returns true if and only if the assertion conditions are met.
-     *
-     * This method is passed the ACL, Role, Resource, and privilege to which the authorization query applies. If the
-     * $role, $calendar, or $privilege parameters are null, it means that the query applies to all Roles, Resources, or
-     * privileges, respectively.
-     *
-     * @param Acl                              $acl
-     * @param RoleInterface                    $role
-     * @param ResourceInterface|CalendarEntity $calendar
-     * @param string                           $privilege
-     *
-     * @return bool
-     */
     public function assert(
         Acl $acl,
         RoleInterface $role = null,
@@ -43,13 +30,15 @@ class Calendar extends AbstractAssertion
         $this->setPrivilege($privilege);
         $id = $id = $this->getId();
 
-        if (!$calendar instanceof CalendarEntity && null !== $id) {
+        if (! $calendar instanceof CalendarEntity && null !== $id) {
             $calendar = $this->calendarService->findCalendarById((int)$id);
         }
 
+        if (! $this->hasContact()) {
+            return false;
+        }
+
         switch ($this->getPrivilege()) {
-            case 'edit':
-                return $this->rolesHaveAccess(Access::ACCESS_OFFICE);
             case 'select-attendees':
                 /**
                  * Stop this case when there is no project calendar
@@ -75,6 +64,8 @@ class Calendar extends AbstractAssertion
                 return true;
             case 'overview-admin':
             case 'view-admin':
+            case 'add-contact':
+            case 'edit':
             case 'edit-attendees-admin':
             case 'set-roles-admin':
             case 'new':
@@ -98,7 +89,7 @@ class Calendar extends AbstractAssertion
                 return $this->rolesHaveAccess($calendar->getType()->getAccess());
 
             case 'view':
-                return $this->calendarService->canViewCalendar($calendar, $this->contact);
+                return $this->calendarService->isPublic($calendar);
         }
 
         return false;
