@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -14,26 +15,26 @@ namespace Calendar\Acl\Assertion;
 
 use Admin\Entity\Access;
 use Calendar\Entity\Document as DocumentEntity;
-use Zend\Permissions\Acl\Acl;
-use Zend\Permissions\Acl\Resource\ResourceInterface;
-use Zend\Permissions\Acl\Role\RoleInterface;
+use Interop\Container\ContainerInterface;
+use Project\Acl\Assertion\Project;
+use Laminas\Permissions\Acl\Acl;
+use Laminas\Permissions\Acl\Resource\ResourceInterface;
+use Laminas\Permissions\Acl\Role\RoleInterface;
 
-class Document extends AbstractAssertion
+final class Document extends AbstractAssertion
 {
     /**
-     * Returns true if and only if the assertion conditions are met.
-     *
-     * This method is passed the ACL, Role, Resource, and privilege to which the authorization query applies. If the
-     * $role, $document, or $privilege parameters are null, it means that the query applies to all Roles, Resources, or
-     * privileges, respectively.
-     *
-     * @param Acl               $acl
-     * @param RoleInterface     $role
-     * @param ResourceInterface $document
-     * @param string            $privilege
-     *
-     * @return bool
+     * @var Project
      */
+    private $projectAssertion;
+
+    public function __construct(ContainerInterface $container)
+    {
+        parent::__construct($container);
+
+        $this->projectAssertion = $container->get(Project::class);
+    }
+
     public function assert(
         Acl $acl,
         RoleInterface $role = null,
@@ -43,13 +44,10 @@ class Document extends AbstractAssertion
         $this->setPrivilege($privilege);
         $id = $this->getId();
 
-        if (!$document instanceof DocumentEntity && null !== $id) {
+        if (! $document instanceof DocumentEntity && null !== $id) {
             $document = $this->calendarService->find(DocumentEntity::class, $id);
         }
 
-        /*
-         * No document was found, so return true because we do not now anything about the access
-         */
         if (null === $document) {
             return true;
         }
@@ -61,7 +59,7 @@ class Document extends AbstractAssertion
                     return true;
                 }
 
-                return $this->calendarService->canViewCalendar($document->getCalendar(), $this->contact);
+                return $this->calendarService->isPublic($document->getCalendar());
             case 'edit-community':
                 if ($this->hasPermission($document->getCalendar(), 'edit')) {
                     return true;
@@ -70,16 +68,18 @@ class Document extends AbstractAssertion
                 /*
                  * The project leader also has rights to invite users
                  */
-                if (null !== $document->getCalendar()->getProjectCalendar()) {
-                    if ($this->contactService->contactHasPermit(
-                        $this->contact,
-                        'edit',
-                        $document->getCalendar()->getProjectCalendar()->getProject()
+                if (
+                    null !== $document->getCalendar()->getProjectCalendar()
+                    && $this->projectAssertion->assert(
+                        $acl,
+                        $role,
+                        $document->getCalendar()->getProjectCalendar()->getProject(),
+                        'edit-community'
                     )
-                    ) {
-                        return true;
-                    }
+                ) {
+                    return true;
                 }
+
 
                 return $this->rolesHaveAccess(Access::ACCESS_OFFICE);
             case 'document-admin':
