@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category   Calendar
  *
  * @author     Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright  Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright  Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license    https://itea3.org/license.txt proprietary
  *
  * @link       https://itea3.org
@@ -38,30 +39,24 @@ class CalendarPdf extends TcpdfFpdi
      */
     public function header()
     {
-        if (\is_null($this->_tplIdx)) {
-            if (!file_exists($this->template)) {
+        if (null === $this->_tplIdx) {
+            if (! file_exists($this->template)) {
                 throw new \InvalidArgumentException(sprintf("Template %s cannot be found", $this->template));
             }
             $this->setSourceFile($this->template);
             $this->_tplIdx = $this->importPage(1);
         }
         $this->SetTopMargin(35);
-        $this->useTemplate($this->_tplIdx, 0, 0);
-
-        //        $this->SetFont('freesans', 'N', 15);
+        $this->useTemplate($this->_tplIdx);
         $this->SetTextColor(0);
         $this->SetXY(15, 5);
     }
 
-    public function footer()
+    public function footer(): void
     {
-        // emtpy method body
     }
 
-    /**
-     * @param $template
-     */
-    public function setTemplate($template)
+    public function setTemplate(string $template): void
     {
         $this->template = $template;
     }
