@@ -1,4 +1,5 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
@@ -12,12 +13,18 @@ declare(strict_types=1);
 
 namespace Calendar\Entity;
 
-use Zend\Permissions\Acl\Resource\ResourceInterface;
+use InvalidArgumentException;
+use Laminas\Permissions\Acl\Resource\ResourceInterface;
+
+use function implode;
+use function sprintf;
+use function str_replace;
+use function strtolower;
 
 /**
  * Class AbstractEntity
  *
- * @package Application\Entity
+ * @package Calendar\Entity
  */
 abstract class AbstractEntity implements EntityInterface, ResourceInterface
 {
@@ -32,33 +39,33 @@ abstract class AbstractEntity implements EntityInterface, ResourceInterface
             case 'class_name':
                 return \str_replace('DoctrineORMModule\Proxy\__CG__\\', '', static::class);
             case 'entity_name':
-                return \implode('', \array_slice(explode('\\', $this->get('class_name')), -1));
+                return implode('', \array_slice(explode('\\', $this->get('class_name')), -1));
             case 'full_entity_name':
-                return \implode('', \explode('\\', $this->get('class_name')));
+                return implode('', \explode('\\', $this->get('class_name')));
             case 'underscore_entity_name':
                 return strtolower(implode('_', \explode('\\', $this->get('class_name'))));
             case 'entity_fieldset_name':
-                return \sprintf(
+                return sprintf(
                     "%sFieldset",
-                    \str_replace('Entity\\', 'Form\\', $this->get('class_name'))
+                    str_replace('Entity\\', 'Form\\', $this->get('class_name'))
                 ); //Run\Form\RunFieldset
             case 'entity_form_name':
-                return \sprintf(
+                return sprintf(
                     "%sForm",
-                    \str_replace('Entity\\', 'Form\\', $this->get('class_name'))
+                    str_replace('Entity\\', 'Form\\', $this->get('class_name'))
                 ); //Run\Form\RunForm
             case 'entity_inputfilter_name':
-                return \sprintf(
+                return sprintf(
                     "%sFilter",
-                    \str_replace('Entity\\', 'InputFilter\\', $this->get('class_name'))
+                    str_replace('Entity\\', 'InputFilter\\', $this->get('class_name'))
                 ); //Run\InputFilter\RunFilter
             case 'entity_assertion_name':
                 return sprintf(
                     "%s",
-                    \str_replace('Entity', 'Acl\\Assertion', $this->get('class_name'))
+                    str_replace('Entity', 'Acl\\Assertion', $this->get('class_name'))
                 ); //Run\Acl\Assertion\Run
             default:
-                throw new \InvalidArgumentException(sprintf("Unknown option %s for get entity name", $switch));
+                throw new InvalidArgumentException(sprintf("Unknown option %s for get entity name", $switch));
         }
     }
 
