@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA copyright message placeholder.
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -14,11 +15,9 @@ namespace Calendar\Entity;
 
 use Doctrine\Common\Collections;
 use Doctrine\ORM\Mapping as ORM;
-use Zend\Form\Annotation;
+use Laminas\Form\Annotation;
 
 /**
- * CalendarContactRole.
- *
  * @ORM\Table(name="calendar_contact_role")
  * @ORM\Entity(repositoryClass="Calendar\Repository\ContactRole")
  */
@@ -28,28 +27,25 @@ class ContactRole extends AbstractEntity
     public const ROLE_STG_REVIEWER = 7;
     public const ROLE_STG_SPARE_REVIEWER = 8;
 
-    /**
-     * @var array Lookup table for the roles
-     */
-    public static $roles
+    public static array $roles
         = [
-            self::ROLE_ATTENDEE           => 'txt-role-attendees',
-            self::ROLE_STG_REVIEWER       => 'txt-stg-reviewer',
+            self::ROLE_ATTENDEE => 'txt-role-attendees',
+            self::ROLE_STG_REVIEWER => 'txt-stg-reviewer',
             self::ROLE_STG_SPARE_REVIEWER => 'txt-stg-spare-reviewer',
         ];
 
     /**
-     * @ORM\Column(name="role_id", type="integer", nullable=false)
+     * @ORM\Column(name="role_id", type="integer", options={"unsigned":true})
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="IDENTITY")
-     * @Annotation\Type("\Zend\Form\Element\Hidden")
+     * @Annotation\Type("\Laminas\Form\Element\Hidden")
      *
-     * @var integer
+     * @var int
      */
     private $id;
     /**
-     * @ORM\Column(name="role", type="string", length=45, nullable=false)
-     * @Annotation\Type("\Zend\Form\Element\Text")
+     * @ORM\Column(name="role", type="string", nullable=false)
+     * @Annotation\Type("\Laminas\Form\Element\Text")
      * @Annotation\Options({"label":"txt-calendar-contact-role-role-label","help-block": "txt-calendar-contact-role-role-help-block"})
      * @Annotation\Attributes({"placeholder":"txt-calendar-contact-role-role-placeholder"})     *
      *
@@ -69,75 +65,52 @@ class ContactRole extends AbstractEntity
         $this->calendarContact = new Collections\ArrayCollection();
     }
 
-    public function __get($property)
-    {
-        return $this->$property;
-    }
-
-    public function __set($property, $value)
-    {
-        $this->$property = $value;
-    }
-
-    public function __isset($property)
-    {
-        return isset($this->$property);
-    }
-
     public function __toString(): string
     {
         return (string)$this->role;
     }
 
-    public function getId()
+    public static function getRoles(): ?array
+    {
+        return self::$roles;
+    }
+
+    /**
+     * @param array $roles
+     */
+    public static function setRoles(array $roles): void
+    {
+        self::$roles = $roles;
+    }
+
+    public function getId(): ?int
     {
         return $this->id;
     }
 
-    /**
-     * @param int $id
-     *
-     * @return ContactRole
-     */
-    public function setId($id): ContactRole
+    public function setId(?int $id): ContactRole
     {
         $this->id = $id;
         return $this;
     }
 
-    /**
-     * @return string
-     */
-    public function getRole()
+    public function getRole(): ?string
     {
         return $this->role;
     }
 
-    /**
-     * @param string $role
-     *
-     * @return ContactRole
-     */
-    public function setRole(string $role): ContactRole
+    public function setRole(?string $role): ContactRole
     {
         $this->role = $role;
         return $this;
     }
 
-    /**
-     * @return Contact[]
-     */
     public function getCalendarContact()
     {
         return $this->calendarContact;
     }
 
-    /**
-     * @param Contact[] $calendarContact
-     *
-     * @return ContactRole
-     */
-    public function setCalendarContact(array $calendarContact): ContactRole
+    public function setCalendarContact($calendarContact): ContactRole
     {
         $this->calendarContact = $calendarContact;
         return $this;
