@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA copyright message placeholder.
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -14,7 +15,7 @@ namespace Calendar\Entity;
 
 use Doctrine\Common\Collections;
 use Doctrine\ORM\Mapping as ORM;
-use Zend\Form\Annotation;
+use Laminas\Form\Annotation;
 
 /**
  * CalendarContactRole.
@@ -28,17 +29,17 @@ class ContactStatus extends AbstractEntity
     public const STATUS_ACCEPT = 2;
     public const STATUS_DECLINE = 3;
     /**
-     * @ORM\Column(name="status_id", type="integer", nullable=false)
+     * @ORM\Column(name="status_id", type="integer", options={"unsigned":true})
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="IDENTITY")
-     * @Annotation\Type("\Zend\Form\Element\Hidden")
+     * @Annotation\Type("\Laminas\Form\Element\Hidden")
      *
-     * @var integer
+     * @var int
      */
     private $id;
     /**
      * @ORM\Column(name="status", type="string", nullable=false)
-     * @Annotation\Type("\Zend\Form\Element\Text")
+     * @Annotation\Type("\Laminas\Form\Element\Text")
      * @Annotation\Options({"label":"txt-calendar-contact-status-status-label","help-block": "txt-calendar-contact-status-status-help-block"})
      * @Annotation\Attributes({"placeholder":"txt-calendar-contact-status-status-placeholder"})
      *
@@ -46,8 +47,8 @@ class ContactStatus extends AbstractEntity
      */
     private $status;
     /**
-     * @ORM\Column(name="status_change", type="string", length=45, nullable=false)
-     * @Annotation\Type("\Zend\Form\Element\Text")
+     * @ORM\Column(name="status_change", type="string", nullable=false)
+     * @Annotation\Type("\Laminas\Form\Element\Text")
      * @Annotation\Options({"label":"txt-calendar-contact-status-status-change-label","help-block": "txt-calendar-contact-status-status-change-help-block"})
      * @Annotation\Attributes({"placeholder":"txt-calendar-contact-status-status-change-placeholder"})
      *
@@ -62,125 +63,54 @@ class ContactStatus extends AbstractEntity
      */
     private $calendarContact;
 
-    /**
-     * Class constructor.
-     */
     public function __construct()
     {
         $this->calendarContact = new Collections\ArrayCollection();
     }
 
-    /**
-     * Magic Getter.
-     *
-     * @param $property
-     *
-     * @return mixed
-     */
-    public function __get($property)
-    {
-        return $this->$property;
-    }
-
-    /**
-     * Magic Setter.
-     *
-     * @param $property
-     * @param $value
-     */
-    public function __set($property, $value)
-    {
-        $this->$property = $value;
-    }
-
-    /**
-     * @return string
-     */
     public function __toString(): string
     {
         return (string)$this->status;
     }
 
-    /**
-     * @param $property
-     *
-     * @return bool
-     */
-    public function __isset($property)
-    {
-        return isset($this->$property);
-    }
-
-    /**
-     * @return int
-     */
-    public function getId()
+    public function getId(): ?int
     {
         return $this->id;
     }
 
-    /**
-     * @param int $id
-     *
-     * @return ContactStatus
-     */
-    public function setId($id): ContactStatus
+    public function setId(?int $id): ContactStatus
     {
         $this->id = $id;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getStatus(): ?string
     {
         return $this->status;
     }
 
-    /**
-     * @param string $status
-     *
-     * @return ContactStatus
-     */
-    public function setStatus(string $status): ContactStatus
+    public function setStatus(?string $status): ContactStatus
     {
         $this->status = $status;
         return $this;
     }
 
-    /**
-     * @return string
-     */
-    public function getStatusChange()
+    public function getStatusChange(): ?string
     {
         return $this->statusChange;
     }
 
-    /**
-     * @param string $statusChange
-     *
-     * @return ContactStatus
-     */
-    public function setStatusChange(string $statusChange): ContactStatus
+    public function setStatusChange(?string $statusChange): ContactStatus
     {
         $this->statusChange = $statusChange;
         return $this;
     }
 
-    /**
-     * @return Contact[]|Collections\ArrayCollection
-     */
     public function getCalendarContact()
     {
         return $this->calendarContact;
     }
 
-    /**
-     * @param Contact[]|Collections\ArrayCollection $calendarContact
-     *
-     * @return ContactStatus
-     */
     public function setCalendarContact($calendarContact): ContactStatus
     {
         $this->calendarContact = $calendarContact;
