@@ -1,4 +1,5 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
@@ -14,11 +15,5 @@ namespace Calendar\Entity;
 
 interface EntityInterface
 {
-    public function __get($property);
-
-    public function __set($property, $value);
-
-    public function __isset($property);
-
     public function getId();
 }
