@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category  Content
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license   https://itea3.org/license.txt proprietary
  *
  * @link      https://itea3.org
@@ -16,24 +17,15 @@ declare(strict_types=1);
 namespace Calendar\Factory;
 
 use Interop\Container\ContainerInterface;
-use Zend\ServiceManager\Factory\FactoryInterface;
+use Laminas\ServiceManager\Factory\FactoryInterface;
 
 /**
- * Factory for instantiating classes with no dependencies or which accept a single array.
+ * Class InvokableFactory
  *
- * The InvokableFactory can be used for any class that:
- *
- * - has no constructor arguments;
- * - accepts a single array of arguments via the constructor.
- *
- * It replaces the "invokables" and "invokable class" functionality of the v2
- * service manager.
+ * @package Calendar\Factory
  */
 final class InvokableFactory implements FactoryInterface
 {
-    /**
-     * {@inheritDoc}
-     */
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
     {
         return (null === $options) ? new $requestedName($container) : new $requestedName($container, $options);
