@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -15,19 +16,20 @@ namespace Calendar\Form;
 use Calendar\Entity\Document;
 use Doctrine\ORM\EntityManager;
 use DoctrineModule\Stdlib\Hydrator\DoctrineObject as DoctrineHydrator;
-use Zend\Form\Form;
-use Zend\InputFilter\InputFilterProviderInterface;
+use Laminas\Form\Form;
+use Laminas\InputFilter\InputFilterProviderInterface;
+use Laminas\Form\Element\Submit;
+use Laminas\Form\Element\Csrf;
+use Laminas\Form\Element\File;
+use Laminas\Form\Element\Text;
 
 /**
+ * Class CreateCalendarDocument
  *
+ * @package Calendar\Form
  */
-class CreateCalendarDocument extends Form implements InputFilterProviderInterface
+final class CreateCalendarDocument extends Form implements InputFilterProviderInterface
 {
-    /**
-     * CreateCalendarDocument constructor.
-     *
-     * @param EntityManager $entityManager
-     */
     public function __construct(EntityManager $entityManager)
     {
         parent::__construct();
@@ -41,70 +43,67 @@ class CreateCalendarDocument extends Form implements InputFilterProviderInterfac
 
         $this->add(
             [
-                'type'       => 'Zend\Form\Element\Text',
+                'type'       => Text::class,
                 'name'       => 'document',
                 'options'    => [
-                    'label'      => _("txt-document-name"),
-                    'help-block' => _("txt-document-name-explanation"),
+                    'label'      => _('txt-document-name'),
+                    'help-block' => _('txt-document-name-explanation'),
                 ],
                 'attributes' => [
                     'class'       => 'form-control',
-                    'placeholder' => _("txt-please-give-a-calendar-document-name"),
+                    'placeholder' => _('txt-please-give-a-calendar-document-name'),
                 ],
             ]
         );
         $this->add(
             [
-                'type'    => '\Zend\Form\Element\File',
+                'type'    => File::class,
                 'name'    => 'file',
                 'options' => [
-                    "label"      => "txt-file",
+                    'label'      => 'txt-file',
                     'class'      => 'form-control',
-                    "help-block" => _("txt-calendar-document-file-help-block"),
+                    'help-block' => _('txt-calendar-document-file-help-block'),
                 ],
             ]
         );
         $this->add(
             [
-                'type' => '\Zend\Form\Element\Csrf',
+                'type' => Csrf::class,
                 'name' => 'csrf',
             ]
         );
         $this->add(
             [
-                'type'       => 'Zend\Form\Element\Submit',
+                'type'       => Submit::class,
                 'name'       => 'submit',
                 'attributes' => [
-                    'class' => "btn btn-primary",
-                    'value' => _("txt-upload-document"),
+                    'class' => 'btn btn-primary',
+                    'value' => _('txt-upload-document'),
                 ],
             ]
         );
         $this->add(
             [
-                'type'       => 'Zend\Form\Element\Submit',
+                'type'       => Submit::class,
                 'name'       => 'delete',
                 'attributes' => [
-                    'class' => "btn btn-danger",
-                    'value' => _("txt-delete"),
+                    'class' => 'btn btn-danger',
+                    'value' => _('txt-delete'),
                 ],
             ]
         );
         $this->add(
             [
-                'type'       => 'Zend\Form\Element\Submit',
+                'type'       => Submit::class,
                 'name'       => 'cancel',
                 'attributes' => [
-                    'class' => "btn btn-warning",
-                    'value' => _("txt-cancel"),
+                    'class' => 'btn btn-warning',
+                    'value' => _('txt-cancel'),
                 ],
             ]
         );
     }
 
-    /**
-     * @return array
-     */
     public function getInputFilterSpecification(): array
     {
         return [
