@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Jield BV all rights reserved.
  *
@@ -15,8 +16,8 @@ namespace Calendar\Form;
 use Calendar\Entity\AbstractEntity;
 use Doctrine\ORM\EntityManager;
 use Interop\Container\ContainerInterface;
-use Zend\Form\Element;
-use Zend\Form\Form;
+use Laminas\Form\Element;
+use Laminas\Form\Form;
 
 /**
  * Class CreateObject
