@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Jield BV all rights reserved.
  *
@@ -22,11 +23,10 @@ use DoctrineModule\Stdlib\Hydrator\DoctrineObject as DoctrineHydrator;
 use DoctrineORMModule\Form\Element\EntityMultiCheckbox;
 use DoctrineORMModule\Form\Element\EntityRadio;
 use DoctrineORMModule\Form\Element\EntitySelect;
-use Zend\Form\Annotation\AnnotationBuilder;
-use Zend\Form\Element;
-use Zend\Form\Element\Radio;
-use Zend\Form\Fieldset;
-use Zend\Form\Form;
+use Laminas\Form\Annotation\AnnotationBuilder;
+use Laminas\Form\Element;
+use Laminas\Form\Element\Radio;
+use Laminas\Form\Fieldset;
 
 /**
  * Class ObjectFieldset
@@ -40,12 +40,6 @@ class ObjectFieldset extends Fieldset
      */
     protected $entityManager;
 
-    /**
-     * @param EntityManager         $entityManager
-     * @param Entity\AbstractEntity $object
-     *
-     * @todo Directly use the form generated by $builder->createForm()
-     */
     public function __construct(EntityManager $entityManager, Entity\AbstractEntity $object)
     {
         parent::__construct($object->get('underscore_entity_name'));
@@ -60,13 +54,6 @@ class ObjectFieldset extends Fieldset
         $this->addElements($data, $object, $this);
     }
 
-    /**
-     * Recursively add form elements and sub-fieldsets
-     *
-     * @param Fieldset|Form         $dataFieldset
-     * @param Entity\AbstractEntity $object
-     * @param Fieldset              $baseFieldset
-     */
     protected function addElements(
         Fieldset $dataFieldset,
         Entity\AbstractEntity $object,
@@ -76,7 +63,7 @@ class ObjectFieldset extends Fieldset
         foreach ($dataFieldset->getElements() as $element) {
             $this->parseElement($element, $object);
             // Add only when a type is provided
-            if (!array_key_exists('type', $element->getAttributes())) {
+            if (! array_key_exists('type', $element->getAttributes())) {
                 continue;
             }
 
@@ -97,15 +84,12 @@ class ObjectFieldset extends Fieldset
         }
     }
 
-    /**
-     * @param Element               $element
-     * @param Entity\AbstractEntity $object
-     */
     protected function parseElement(Element $element, Entity\AbstractEntity $object): void
     {
         // Go over each element to add the objectManager to the EntitySelect
         /** Element $element */
-        if ($element instanceof EntitySelect || $element instanceof EntityMultiCheckbox
+        if (
+            $element instanceof EntitySelect || $element instanceof EntityMultiCheckbox
             || $element instanceof EntityRadio
         ) {
             $element->setOptions(
@@ -115,7 +99,7 @@ class ObjectFieldset extends Fieldset
                 )
             );
         }
-        if ($element instanceof Radio && !$element instanceof EntityRadio) {
+        if ($element instanceof Radio && ! $element instanceof EntityRadio) {
             $attributes = $element->getAttributes();
             $valueOptionsArray = sprintf('get%s', ucfirst($attributes['array']));
             $element->setOptions(
