@@ -1,13 +1,9 @@
 <?php
+
 /**
- * ITEA Office all rights reserved
- *
- * PHP Version 7
- *
- * @category    Project
- *
+*
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright   Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license     https://itea3.org/license.txt proprietary
  *
  * @link        http://github.com/iteaoffice/project for the canonical source repository
@@ -17,19 +13,16 @@ declare(strict_types=1);
 
 namespace Calendar\InputFilter;
 
-use Zend\InputFilter\InputFilter;
-use Zend\Validator\Callback;
+use Laminas\InputFilter\InputFilter;
+use Laminas\Validator\Callback;
 
 /**
  * Class CalendarFilter
  *
  * @package Calendar\InputFilter
  */
-class CalendarFilter extends InputFilter
+final class CalendarFilter extends InputFilter
 {
-    /**
-     * CalendarFilter constructor.
-     */
     public function __construct()
     {
         $inputFilter = new InputFilter();
