@@ -1,13 +1,9 @@
 <?php
+
 /**
- * ITEA Office all rights reserved
- *
- * PHP Version 7
- *
- * @category    Project
  *
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright   Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license     https://itea3.org/license.txt proprietary
  *
  * @link        http://github.com/iteaoffice/project for the canonical source repository
@@ -17,20 +13,17 @@ declare(strict_types=1);
 
 namespace Calendar\InputFilter;
 
-use Zend\InputFilter\FileInput;
-use Zend\InputFilter\InputFilter;
-use Zend\Validator\File\Size;
+use Laminas\InputFilter\FileInput;
+use Laminas\InputFilter\InputFilter;
+use Laminas\Validator\File\Size;
 
 /**
  * Class DocumentFilter
  *
  * @package Calendar\InputFilter
  */
-class DocumentFilter extends InputFilter
+final class DocumentFilter extends InputFilter
 {
-    /**
-     * DocumentFilter constructor.
-     */
     public function __construct()
     {
         $inputFilter = new InputFilter();
