@@ -1,13 +1,9 @@
 <?php
+
 /**
- * ITEA Office all rights reserved
- *
- * PHP Version 7
- *
- * @category    Project
  *
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright   Copyright (c) 2019 ITEA Office (https://itea3.org)
  * @license     https://itea3.org/license.txt proprietary
  *
  * @link        http://github.com/iteaoffice/project for the canonical source repository
@@ -19,26 +15,22 @@ namespace Calendar\Navigation\Invokable;
 
 use Admin\Navigation\Invokable\AbstractNavigationInvokable;
 use Calendar\Entity\Calendar;
-use Zend\Navigation\Page\Mvc;
+use Laminas\Navigation\Page\Mvc;
 
 /**
  * Class ProjectLabel
  *
  * @package Project\Navigation\Invokable
  */
-class CalendarLabel extends AbstractNavigationInvokable
+final class CalendarLabel extends AbstractNavigationInvokable
 {
-    /**
-     * @param Mvc $page
-     *
-     * @return void;
-     */
     public function __invoke(Mvc $page): void
     {
+        $label = $this->translate('txt-nav-calendar');
         if ($this->getEntities()->containsKey(Calendar::class)) {
+
             /** @var Calendar $calendar */
             $calendar = $this->getEntities()->get(Calendar::class);
-
 
             $page->setParams(
                 array_merge(
@@ -49,8 +41,6 @@ class CalendarLabel extends AbstractNavigationInvokable
                 )
             );
             $label = (string)$calendar;
-        } else {
-            $label = $this->translate('txt-nav-calendar');
         }
         $page->set('label', $label);
     }
