@@ -1,11 +1,12 @@
 <?php
+
 /**
  * ITEA Office all rights reserved
  *
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -13,8 +14,9 @@ declare(strict_types=1);
 namespace Calendar\Repository;
 
 use Calendar\Entity;
-use Calendar\Service\CalendarService;
 use Contact\Entity\Contact as ContactEntity;
+use DateTime;
+use Doctrine\Common\Collections\Criteria;
 use Doctrine\ORM\EntityRepository;
 
 /**
@@ -24,7 +26,7 @@ use Doctrine\ORM\EntityRepository;
  */
 final class Contact extends EntityRepository
 {
-    public function findCalendarContactByContact(string $which, ContactEntity $contact): array
+    public function findUpcomingCalendarContactByContact(ContactEntity $contact): array
     {
         $qb = $this->_em->createQueryBuilder();
         $qb->select('calendar_entity_contact');
@@ -32,46 +34,28 @@ final class Contact extends EntityRepository
         $qb->join('calendar_entity_contact.calendar', 'calendar_entity_calendar');
         $qb->join('calendar_entity_contact.contact', 'contact_entity_contact');
 
-        switch ($which) {
-            case CalendarService::WHICH_UPCOMING:
-                $qb->andWhere('calendar_entity_calendar.dateFrom >= ?1');
-                $qb->orderBy('calendar_entity_calendar.dateFrom', 'ASC');
-                $qb->setParameter(1, new \DateTime());
-                break;
-            case CalendarService::WHICH_PAST:
-                $qb->andWhere('calendar_entity_calendar.dateFrom <= ?1');
-                $qb->orderBy('calendar_entity_calendar.dateEnd', 'DESC');
-                $qb->setParameter(1, new \DateTime());
-                break;
-            case CalendarService::WHICH_REVIEWS:
-                $qb->andWhere('calendar_entity_calendar.dateEnd >= ?1');
-                $qb->orderBy('calendar_entity_calendar.dateFrom', 'ASC');
-                $qb->setParameter(1, new \DateTime());
-                $projectCalendarSubSelect = $this->_em->createQueryBuilder();
-                $projectCalendarSubSelect->select('calendar.id');
-                $projectCalendarSubSelect->from(\Project\Entity\Calendar\Calendar::class, 'projectCalendar');
-                $projectCalendarSubSelect->join('projectCalendar.calendar', 'calendar');
-                $qb->andWhere($qb->expr()->in('calendar_entity_calendar.id', $projectCalendarSubSelect->getDQL()));
-                break;
-            case CalendarService::WHICH_UPDATED:
-                $qb->orderBy('calendar_entity_calendar.dateUpdated', 'DESC');
-                break;
-            case CalendarService::WHICH_ON_HOMEPAGE:
-                $qb->andWhere('calendar_entity_calendar.dateEnd >= ?1');
-                $qb->setParameter(1, new \DateTime());
-                $qb->andWhere('calendar_entity_calendar.onHomepage = ?2');
-                $qb->setParameter(2, Entity\Calendar::ON_HOMEPAGE);
-                $qb->andWhere('calendar_entity_calendar.final = ?3');
-                $qb->setParameter(3, Entity\Calendar::FINAL_FINAL);
-                $qb->addOrderBy('calendar_entity_calendar.dateFrom', 'ASC');
-                $qb->orderBy('calendar_entity_calendar.sequence', 'ASC');
+        $qb->andWhere('calendar_entity_contact.contact = :contact');
+        $qb->andWhere('calendar_entity_calendar.dateEnd >= :today');
+        $qb->setParameter('today', new DateTime());
+        $qb->addOrderBy('calendar_entity_calendar.dateFrom', Criteria::ASC);
 
-                break;
-        }
+        $qb->setParameter('contact', $contact);
 
-        $qb->andWhere('calendar_entity_contact.contact = ?10');
-        $qb->addOrderBy('contact_entity_contact.lastName', 'ASC');
-        $qb->setParameter(10, $contact);
+        return $qb->getQuery()->getResult();
+    }
+
+    public function findCalendarContactByContact(ContactEntity $contact): array
+    {
+        $qb = $this->_em->createQueryBuilder();
+        $qb->select('calendar_entity_contact');
+        $qb->from(Entity\Contact::class, 'calendar_entity_contact');
+        $qb->join('calendar_entity_contact.calendar', 'calendar_entity_calendar');
+        $qb->join('calendar_entity_contact.contact', 'contact_entity_contact');
+
+        $qb->andWhere('calendar_entity_contact.contact = :contact');
+        $qb->addOrderBy('calendar_entity_calendar.dateFrom', Criteria::ASC);
+
+        $qb->setParameter('contact', $contact);
 
         return $qb->getQuery()->getResult();
     }
