@@ -5,7 +5,7 @@
  * @category  Calendar
  *
  * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 declare(strict_types=1);
@@ -17,5 +17,5 @@ namespace Calendar\Version;
  */
 final class Version
 {
-    public const VERSION = '4.0.0';
+    public const VERSION = '5.0.0';
 }
