@@ -5,20 +5,17 @@
  * @category    CalendarTest
  * @package     Entity
  * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @copyright   Copyright (c) 2019 ITEA Office (https://itea3.org)
  */
 
 namespace CalendarTest\Entity;
 
 use PHPUnit\Framework\TestCase;
-use Zend\Form\Annotation\AnnotationBuilder;
-use Zend\Form\Element;
+use Laminas\Form\Annotation\AnnotationBuilder;
+use Laminas\Form\Element;
 
 class EntityTest extends TestCase
 {
-    /**
-     *
-     */
     public function testCanCreateEntitiesAndSaveTxtFields(): void
     {
         $labels = [];
@@ -31,9 +28,8 @@ class EntityTest extends TestCase
 
             $testClass = new \ReflectionClass($className);
 
-
             if ($testClass->isInstantiable()) {
-                $object = new $className;
+                $object = new $className();
 
                 $this->assertInstanceOf($className, $object);
 
@@ -67,8 +63,8 @@ class EntityTest extends TestCase
                         $labels[] = $element->getOptions()['placeholder'];
                     }
 
-                    $this->assertInternalType('array', ($element->getAttributes()));
-                    $this->assertInternalType('array', ($element->getOptions()));
+                    $this->assertIsArray($element->getAttributes());
+                    $this->assertIsArray($element->getOptions());
                 }
 
                 foreach ($testClass->getStaticProperties() as $constant) {
@@ -83,7 +79,7 @@ class EntityTest extends TestCase
 
         file_put_contents(
             __DIR__ . '/../../config/language.php',
-            "<?php\n_(\"" . implode("\");\n_(\"", array_unique($labels)) . '");'
+            "<?php\n_('" . implode("');\n_('", array_unique($labels)) . "');\n"
         );
     }
 }
